Fix Unicode handling.
I'm not sure what I was thinking, but I was computing the number
of display columns by counting the number of bytes used in each
code point.

I've fixed this so that if the bytes make up a Unicode string, then
we try to compute the display columns used. Otherwise, we fall back
to the number of bytes as the number of columns.