a trailing-cell in right and center alignment mode
Previously, a **cell** was a block of text to the left of some tab.

This meant that a table like this:

   aa\tbb
   a\tb

would be formatted as

   aa bb
    a b

in right-alignment mode because the b's were not actually in a cell, and
thus didn't get aligned.

This commit modifies the notion of a "cell" in right- and
center-alignment mode to also include text to the right of the last tab
on the line.

This text is now considered part of a cell and aligned accordingly.

Implemented as describe above, this change would cause many trailing
spaces to be printed in right- and center-alignment modes, including on
empty lines.

To alleviate this, trailing spaces in the final cell of a line are not
printed.

fixes #29