api: add TabWriter::ansi option and drop regex crate dependency
This finally does the work to just remove the `regex` dependency
entirely. The `regex` dependency was the primary reason that the ANSI
handling was a crate feature. Namely, it permitted callers who didn't
care about ANSI handling to avoid depending on the `regex` crate. But
now that we drop the use of the `regex` crate entirely, we can just make
ANSI handling a normal option of `TabWriter`. And that's what we do. We
disable it by default.

This also deprecates the `ansi_formatting` feature. Now if one uses it,
the only effect it has is to enable the `TabWriter::ansi` option by
default. This avoids any breaking changes.

(I did consider replacing `regex` with `regex-lite`, which would have
been an improvement. Especially since we were already disabling most of
the regex crate's features anyway. But the regex was simple enough to
just hand-code.)

We also bump to Rust 2021.